refactor: ai 패키지 분리
